+++ react/modules/sample/src/components/SampleCard.js
@@ -15,10 +15,6 @@ const SampleCard = () => {
         link: `/${window?.contextPath}/employee/sample/create`,
       },
       {
-        label: t("Sample Search"),
-        link: `/${window?.contextPath}/employee/sample/search`,
-      },
-      {
         label: t("New Sample Search"),
         link: `/${window?.contextPath}/employee/sample/new-search`,
       },
@@ -54,6 +50,14 @@ const SampleCard = () => {
         label: t("Boundary with SearchComposerWithFilter"),
         link: `/${window?.contextPath}/employee/sample/searchComposerWithFilter`,
       },
+      {
+        label: t("Sample Search"),
+        link: `/${window?.contextPath}/employee/sample/search`,
+      },
+      {
+        label: t("Sample Search with Multiple Tabs"),
+        link: `/${window?.contextPath}/employee/sample/search-multiple-tabs`,
+      }
     ],
   };
   return <EmployeeModuleCard {...propsForModuleCard} />;